spellfile updates

+++ .vim/spell/spellfile_update.sh
@@ -8,13 +8,14 @@ customPath="${externalDir}/spellfile_custom.txt"
 externalPath="${spellDir}/external.add"
 gboardPath="${externalDir}/gboard_spellfile.txt"
 
-# TODO dedupe core spellfiles
-
 # unzip any zips in externalDir
-if 7z e "${externalDir}"/*.zip -aoa -o"${externalDir}"
+if ls "${externalDir}"/*.zip 1> /dev/null 2>&1
 then
-	# burn zips on success
-	rm "${externalDir}"/*.zip
+	if 7z e "${externalDir}"/*.zip -aoa -o"${externalDir}"
+	then
+		# burn zips on success
+		rm "${externalDir}"/*.zip
+	fi
 fi
 
 # find new words from any text files in externalDir
@@ -44,3 +45,16 @@ newGboardText="# Gboard Dictionary version:1
 $(sed -e 's/$/\ten-US/g' -e 's/^/\t/g' ${customPath})"
 echo "${newGboardText}" > "${gboardPath}"
 
+# TODO: actually dedupe files?
+echo "checking for dupes..."
+spellFiles="$(find "$spellDir" -type f -iname '*.add')"
+echo "$spellFiles" | while read -r src
+do
+	echo "$spellFiles" | while read -r check
+	do
+		# 2>/dev/null isn't working for eating errors
+		# so piping to echo -n instead
+		# could sort with 'sort -t ':' -k2', but would need to aggregate output first
+		grep --fixed-strings --line-regexp --color --with-filename --exclude "$src" -f "$src" "$check" || echo -n
+	done
+done
